chore: Move MicroFrontendProxyProvider to turborepo-task-executor
- Move MicroFrontendProxyProvider struct to turborepo-task-executor/src/command.rs
- Make it generic over MfeConfigProvider trait instead of concrete MicrofrontendsConfigs
- Add RecursiveTurboError to new error.rs file in turborepo-task-executor
- Extend MfeConfigProvider trait with dev_tasks() and config_filename() methods
- Implement new trait methods on MicrofrontendsConfigs in turborepo-lib
- Simplify turborepo-lib visitor/command.rs to type aliases and tests

This continues the turborepo-lib modularization effort, reducing coupling
and improving testability of task execution components.
